feat: add new angular signals api
