+++ packages/core/src/generators/angular/helpers/format.ts
@@ -1,0 +1,20 @@
+import { format } from 'prettier/standalone';
+
+export const tryFormat = (str: string, parser: string) => {
+  try {
+    return format(str, {
+      parser,
+      plugins: [
+        // To support running in browsers
+        require('prettier/parser-typescript'),
+        require('prettier/parser-postcss'),
+        require('prettier/parser-html'),
+        require('prettier/parser-babel'),
+      ],
+      htmlWhitespaceSensitivity: 'ignore',
+    });
+  } catch (err) {
+    console.warn('Could not prettify', { string: str }, err);
+  }
+  return str;
+};
