+++ packages/core/src/generators/angular/helpers/parse-selector.ts
@@ -1,6 +1,6 @@
 import { CssSelector } from '@angular/compiler';
 
-export function parse(selector: string) {
+export function parseSelector(selector: string) {
   const { element, classNames, attrs } = CssSelector.parse(selector)[0];
   const attributes = attrs.reduce((acc, attr, i) => {
     if (i % 2 === 0) {
